% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sort_annotations}
\alias{sort_annotations}
\title{Sort a GTF/GFF/GFF3 file.}
\usage{
sort_annotations(input, output, threads)
}
\arguments{
\item{input}{The input file path.}

\item{output}{The output file path.}

\item{threads}{The number of threads to use.}
}
\value{
a list with the input and output file paths, the number of threads used, whether the input and output were memory-mapped, the time taken to parse, index, and write the output, and the memory used before and after the operation.
}
\description{
Sort a GTF/GFF/GFF3 file.
}
\examples{
sort_annotations("tests/data/chr1.gtf", "tests/data/chr1.sorted.gtf", 1)

}
