% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sort_annotations_str}
\alias{sort_annotations_str}
\title{Sort a string with GTF/GFF/GFF3 annotations.}
\usage{
sort_annotations_str(mode, input, output, threads)
}
\arguments{
\item{mode}{The mode to parse the annotations. Either "gtf" or "gff" or "gff3".}

\item{input}{The string with the GTF/GFF/GFF3 annotations.}

\item{output}{A function that will be called with each chunk of the sorted string. Return NULL to continue, or a string to stop.}

\item{threads}{The number of threads to use.}
}
\value{
a list with the input and output strings, the number of threads used, whether the input and output were memory-mapped, the time taken to parse, index, and write the output, and the memory used before and after the operation.
}
\description{
Sort a string with GTF/GFF/GFF3 annotations.
}
\examples{
sort_annotations_str("gtf", "chr1\t.\texon\t11869\t12227\t.\t+\t.\tgene_id \"ENSG00000223972.5\"; transcript_id \"ENST00000456328.2\"; exon_number \"1\";\nchr1\t.\texon\t12613\t12721\t.\t+\t.\tgene_id \"ENSG00000223972.5\"; transcript_id \"ENST00000456328.2\"; exon_number \"2\";", function(str) { cat(str); return(NULL); }, 1)

}
